Modularize input processing
